# src/Services/Scheduling/Api/openapi/Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "KDVManager Scheduling API",
    "contact": {
      "name": "Luuk van Hulten",
      "email": "admin@kdvmanager.nl"
    },
    "version": "v1"
  },
  "paths": {
...
    "/v1/groups": {
      "get": {
        "tags": [
          "Groups"
        ],
        "operationId": "ListGroups",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupListVM"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupListVM"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupListVM"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "operationId": "AddGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddGroupCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddGroupCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddGroupCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/schedules": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "operationId": "GetChildSchedules",
        "parameters": [
          {
            "name": "childId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChildScheduleListVM"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChildScheduleListVM"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChildScheduleListVM"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Schedules"
        ],
        "operationId": "AddSchedule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddScheduleCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddScheduleCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddScheduleCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/schedules/daterange": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "operationId": "GetSchedulesByDate",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "groupId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleByDateVM"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleByDateVM"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleByDateVM"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/schedules/group-summary": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "operationId": "GetGroupSummary",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GroupSummaryVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupSummaryVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupSummaryVM"
                }
              }
            }
          }
        }
      }
    },
    "/v1/schedules/print": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "operationId": "GetPrintSchedules",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "groupId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "groupIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PrintSchedulesVM"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintSchedulesVM"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintSchedulesVM"
                }
              }
            }
          }
        }
      }
    },
    "/v1/schedules/{id}": {
      "put": {
        "tags": [
          "Schedules"
        ],
        "operationId": "UpdateSchedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessableEntityResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Schedules"
        ],
        "operationId": "DeleteSchedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddGroupCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "AddScheduleCommand": {
        "type": "object",
        "properties": {
          "childId": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "scheduleRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddScheduleCommandScheduleRule"
            }
          }
        }
      },
      "AddScheduleCommandScheduleRule": {
        "type": "object",
        "properties": {
          "day": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "timeSlotId": {
            "type": "string",
            "format": "uuid"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "AgeGroupSummary": {
        "type": "object",
        "properties": {
          "ageRange": {
            "type": "string"
          },
          "childCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BkrAgeRatio": {
        "required": [
          "age",
          "maxChildrenPerProfessional"
        ],
        "type": "object",
        "properties": {
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "maxChildrenPerProfessional": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BkrAppliedRule": {
        "required": [
          "minAge",
          "maxAge",
          "maxChildren",
          "minProfessionals",
          "constraints",
          "ratios"
        ],
        "type": "object",
        "properties": {
          "minAge": {
            "type": "integer",
            "format": "int32"
          },
          "maxAge": {
            "type": "integer",
            "format": "int32"
          },
          "maxChildren": {
            "type": "integer",
            "format": "int32"
          },
          "minProfessionals": {
            "type": "integer",
            "format": "int32"
          },
          "constraints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BkrRuleConstraint"
            }
          },
          "ratios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BkrAgeRatio"
            }
          }
        }
      },
      "BkrCalculationSummary": {
        "required": [
          "hasSolution",
          "basis"
        ],
        "type": "object",
        "properties": {
          "hasSolution": {
            "type": "boolean"
          },
          "basis": {
            "$ref": "#/components/schemas/BkrProfessionalsBasis"
          },
          "appliedRule": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BkrAppliedRule"
              }
            ]
          }
        }
      },
      "BkrProfessionalsBasis": {
        "enum": [
          "None",
          "GroupSizeMinimum",
          "RatioCalculation",
          "OneChildLessSafeguard"
        ]
      },
      "BkrRuleConstraint": {
        "required": [
          "minAge",
          "maxAge",
          "maxChildren"
        ],
        "type": "object",
        "properties": {
          "minAge": {
            "type": "integer",
            "format": "int32"
          },
          "maxAge": {
            "type": "integer",
            "format": "int32"
          },
          "maxChildren": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ChildScheduleListVM": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "childId": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "scheduleRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChildScheduleListVMScheduleRule"
            }
          }
        }
      },
      "ChildScheduleListVMScheduleRule": {
        "type": "object",
        "properties": {
          "day": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "timeSlotId": {
            "type": "string",
            "format": "uuid"
          },
          "timeSlotName": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "time"
          },
          "endTime": {
            "type": "string",
            "format": "time"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "groupName": {
            "type": "string"
          }
        }
      },
...
      "DayOfWeek": {
        "type": "integer"
      },
...
      "GroupListVM": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "GroupSummaryVM": {
        "required": [
          "groupId",
          "groupName",
          "date",
          "timeBlocks"
        ],
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "groupName": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "timeBlocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeBlockSummary"
            }
          },
          "requiredProfessionals": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfChildren": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PrintCellVM": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "absenceType": {
            "type": [
              "null",
              "string"
            ]
          },
          "startTime": {
            "type": [
              "null",
              "string"
            ]
          },
          "endTime": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PrintChildVM": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "givenName": {
            "type": "string"
          },
          "familyName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "schedule": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/PrintCellVM"
            }
          },
          "index": {
            "type": "integer",
            "format": "int32"
          },
          "ageDisplay": {
            "type": "string"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PrintGroupVM": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrintGroupWeekdayPageVM"
            }
          }
        }
      },
      "PrintGroupWeekdayPageVM": {
        "type": "object",
        "properties": {
          "weekday": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dates": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date"
            }
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrintChildVM"
            }
          }
        }
      },
      "PrintSchedulesVM": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrintGroupVM"
            }
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ScheduleByDateVM": {
        "type": "object",
        "properties": {
          "scheduleId": {
            "type": "string",
            "format": "uuid"
          },
          "childId": {
            "type": "string",
            "format": "uuid"
          },
          "childFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "timeSlotName": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "time"
          },
          "endTime": {
            "type": "string",
            "format": "time"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "dateOfBirth": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TimeBlockSummary": {
        "required": [
          "startTime",
          "endTime",
          "timeSlotName",
          "totalChildren",
          "ageGroups",
          "bkr"
        ],
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "format": "time"
          },
          "endTime": {
            "type": "string",
            "format": "time"
          },
          "timeSlotName": {
            "type": "string"
          },
          "totalChildren": {
            "type": "integer",
            "format": "int32"
          },
          "requiredProfessionals": {
            "type": "integer",
            "format": "int32"
          },
          "ageGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgeGroupSummary"
            }
          },
          "bkr": {
            "$ref": "#/components/schemas/BkrCalculationSummary"
          }
        }
      },
...
      "UnprocessableEntityResponse": {
        "required": [
          "status",
          "errors"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "UpdateScheduleCommand": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "scheduleRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateScheduleCommandScheduleRule"
            }
          }
        }
      },
      "UpdateScheduleCommandScheduleRule": {
        "type": "object",
        "properties": {
          "day": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "timeSlotId": {
            "type": "string",
            "format": "uuid"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ValidationError": {
        "required": [
          "property",
          "code",
          "title"
        ],
        "type": "object",
        "properties": {
          "property": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      }
    }
  },
...
}